[mutable] Fix performance loss on background image loading when editor is not focused. Fix bugs related to using floats for platform time. [REVIEW] [at]gerard.martin #rnx
[CL 28035752 by jordi rovira in ue5-main branch]

+++ Engine/Plugins/Experimental/Mutable/Source/CustomizableObject/Private/MuCO/FMutableTaskGraph.cpp
@@ -142,8 +142,8 @@ void FMutableTaskGraph::TryLaunchMutableTaskLowPriority(bool bFromMutableTask)
 		return;
 	}
 
-	float TimeLimit;
-	float TimeElapsed;
+	double TimeLimit;
+	double TimeElapsed;
 	
 	bool bTimeLimit;
